+++ agb/src/display/object/sprites.rs
@@ -6,6 +6,5 @@ const BYTES_PER_TILE_8BPP: usize = 16;
 
 pub use sprite::{PaletteMulti, Size, Sprite, Tag, include_aseprite};
 pub use sprite_allocator::{
-    DynamicSprite16, DynamicSprite256, PaletteVram, PaletteVramMulti, PaletteVramSingle,
-    SPRITE_LOADER, SpriteVram,
+    DynamicSprite16, DynamicSprite256, PaletteVram, PaletteVramMulti, PaletteVramSingle, SpriteVram,
 };